feat: webpack v5
BREAKING CHANGE: This version adds support to `webpack` `v5` and
drops `webpack` `v4`. Use a previous version if using `webpack` `v4`.

DEPRECATED: This version drops the options `debugLogs` and `silent`,
as it now uses webpack's built-in logging system. Logging should be
enabled through webpack's own configuration. The option
`debugLifecycle` is still present, but it also works in conjunction
with webpack's logging system, so both need to be enabled for logs to
 be seen.